feat: Add about and blog pages
This commit adds the about and blog pages to the project. The about page contains a simple layout with a heading, paragraph, and a link to the home page. The blog page also has a similar layout with a heading, paragraph, and links to the home page and the about page.